refactor(mis): move config to libs/config
